requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.19
groq==0.4.2
python-dotenv==1.0.0

# Search APIs
google-search-results==2.4.2

# Web scraping and content extraction
trafilatura==1.6.1
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0

# Frontend
streamlit==1.30.0

# Utilities
typing-extensions>=4.5.0
